refactor: useChatbot
